ex new first magical problem
